requirements.txt
# ============================================================================
# INTENTIONALLY VULNERABLE REQUIREMENTS - FOR SECURITY SCANNER TESTING ONLY
# These package versions have KNOWN CVEs to trigger Trivy/dependency scanners.
# DO NOT install these in production environments.
# ============================================================================
Django==3.1
Flask==1.0
requests==2.25.0
Jinja2==2.11.3
cryptography==3.4
Pillow==8.0
urllib3==1.26.4
paramiko==2.7.2
PyYAML==5.3.1
lxml==4.6.2
numpy==1.19.4
setuptools==49.6.0
pip==21.0
certifi==2020.12.5
